+++ R/parsnip-temporal_hierarchy_data.R
@@ -6,21 +6,21 @@
 # nocov start
 
 
-make_hierarchical_reg <- function() {
+make_temporal_hierarchy <- function() {
 
-    parsnip::set_new_model("hierarchical_reg")
-    parsnip::set_model_mode("hierarchical_reg", "regression")
+    parsnip::set_new_model("temporal_hierarchy")
+    parsnip::set_model_mode("temporal_hierarchy", "regression")
 
-    # HIERARCHICAL_REG ----
+    # temporal_hierarchy ----
 
     # * Model ----
-    parsnip::set_model_engine("hierarchical_reg", mode = "regression", eng = "thief")
-    parsnip::set_dependency("hierarchical_reg", "thief", "thief")
-    parsnip::set_dependency("hierarchical_reg", "thief", "modeltime")
+    parsnip::set_model_engine("temporal_hierarchy", mode = "regression", eng = "thief")
+    parsnip::set_dependency("temporal_hierarchy", "thief", "thief")
+    parsnip::set_dependency("temporal_hierarchy", "thief", "modeltime")
 
     # * Args ----
     parsnip::set_model_arg(
-        model        = "hierarchical_reg",
+        model        = "temporal_hierarchy",
         eng          = "thief",
         parsnip      = "combination_method",
         original     = "comb",
@@ -29,7 +29,7 @@ make_hierarchical_reg <- function() {
     )
 
     parsnip::set_model_arg(
-        model        = "hierarchical_reg",
+        model        = "temporal_hierarchy",
         eng          = "thief",
         parsnip      = "use_model",
         original     = "usemodel",
@@ -40,7 +40,7 @@ make_hierarchical_reg <- function() {
 
     # * Encoding ----
     parsnip::set_encoding(
-        model   = "hierarchical_reg",
+        model   = "temporal_hierarchy",
         eng     = "thief",
         mode    = "regression",
         options = list(
@@ -53,20 +53,20 @@ make_hierarchical_reg <- function() {
 
     # * Fit ----
     parsnip::set_fit(
-        model         = "hierarchical_reg",
+        model         = "temporal_hierarchy",
         eng           = "thief",
         mode          = "regression",
         value         = list(
             interface = "data.frame",
             protect   = c("x", "y"),
-            func      = c(fun = "hierarchical_fit_impl"),
+            func      = c(fun = "temporal_hier_fit_impl"),
             defaults  = list()
         )
     )
 
     # * Predict ----
     parsnip::set_pred(
-        model         = "hierarchical_reg",
+        model         = "temporal_hierarchy",
         eng           = "thief",
         mode          = "regression",
         type          = "numeric",